start writing a filer based on tags
having some issues with scopes. Should the entire view be in controller? Probably not.
Okay, how do I share share the list of selected tags with the filter function then?
Maybe replace the tag.selected binding with a method call. The method would be a method call to a service
use ng-change on the "filter"