update geolocator to work even if multiple directives are loaded at once
Previously multiple directive instances loaded the Google Maps API for
for every instance. Also, the only one of the instances got notified of the
user location, others were left pending.

Also, timeout was added, so that if the user doesn't approve the usage of
their location within 10 seconds, an error message can be displayed.