Order selected tags into the front
This way it's better visible which tags are and which aren't chosen,
even if there's more tags then currently shown. It could still be an
issue if there's more than 5 (the number of tags shown by default) tags
selected, as some of the selected ones can go out of the view if the
list is limited. A good solution for this is TBD. One idea would be to
always show at least as many items as are selected.